+++ packages/shared/src/index.ts
@@ -88,3 +88,4 @@ export * from "./helpdesk.js";
 export * from "./appBuilder.js";
 export * from "./businessIntelligence.js";
 export * from "./enterpriseSearch.js";
+export * from "./softwareFactory.js";